requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alibi-detect==0.13.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bertopic==0.17.4
blinker==1.9.0
cachetools==7.0.6
catalogue==2.0.10
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.5
delta_spark==4.2.0
dill==0.3.9
filelock==3.29.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.48
h11==0.16.0
hdbscan==0.8.42
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.13
ImageIO==2.37.3
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.1
kiwisolver==1.5.0
lazy-loader==0.5
llvmlite==0.42.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.20.0
networkx==3.6.1
nltk==3.9.4
numba==0.59.1
numpy==1.26.4
opencv-python==4.11.0.86
packaging==26.2
pandas==2.3.3
pillow==10.4.0
plotly==6.7.0
POT==0.9.6.post1
propcache==0.4.1
protobuf==7.34.1
py4j==0.10.9.9
pyarrow==24.0.0
pydantic==2.13.4
pydantic_core==2.46.4
pydeck==0.9.2
Pygments==2.20.0
pynndescent==0.6.0
pyparsing==3.3.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
pytz==2026.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
textblob==0.20.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
umap-learn==0.5.12
urllib3==2.6.3
xxhash==3.7.0
yarl==1.23.0
zipp==3.23.1
